# server/swagger.yaml
swagger: '2.0'
info:
  description: |
    This api is consumed by our frontend react app in order to keep track of the whitelist requests for the game server
  version: 1.0.0
  title: mc-whitelist API
  termsOfService: http://swagger.io/terms/
  contact:
    email: tiaven1104@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# basePath: /api/v1
...
paths:
  /requests/:
    post:
      tags:
      - requests
      summary: Create a new whitelist request
      operationId: createRequest
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Whitelist object that needs to be created
        required: true
        schema:
          $ref: '#/definitions/CreateRequest'
      responses:
        400:
          description: Invalid request body
        500:
          description: Internal server error
        422:
          description: There is a pending request associated with this username
        409:
          description: The request associated with this username is already approved
        201:
          description: Request created
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  CreateRequest:
    type: object
    required:
    - username
    - email
    - age
    - gender
    - info
    properties:
      info:
        $ref: '#/definitions/Info'
      username:
        type: string
        example: doggie
      email:
        type: string
        example: doggie@gmail.com
      age:
        type: integer
        format: int32
        example: 19
      gender:
        type: string
        enum:
        - male
        - female
        - other
        example: female
  Info:
    type: object
    required: 
    - applicationText
    properties:
      applicationText:
        type: string
        example: I'd like to join the server
...
host: virtserver.swaggerhub.com
basePath: /tywinzhang/mc-whitelist/1.0.0
schemes:
 - https